Extracted ValidatorFutureCommittees to a separate component
