+++ packages/page-staking/src/Query/index.tsx
@@ -3,21 +3,17 @@
 
 import type { SessionIndex } from '@polkadot/types/interfaces';
 import type { INumber } from '@polkadot/types/types';
-import type { FutureCommittee } from '../Performance/useFutureSessionCommittee.js';
 
-import React, { useCallback, useMemo, useRef, useState } from 'react';
+import React, { useCallback, useMemo, useState } from 'react';
 import { useParams } from 'react-router-dom';
 
 import { getCommitteeManagement } from '@polkadot/react-api/getCommitteeManagement';
-import { Button, CardSummary, InputAddressSimple, Spinner, styled, SummaryBox, Table, ToggleGroup } from '@polkadot/react-components';
-import { useApi, useCall, useNextTick } from '@polkadot/react-hooks';
+import { Button, CardSummary, InputAddressSimple, Spinner, styled, SummaryBox, ToggleGroup } from '@polkadot/react-components';
+import { useApi, useCall } from '@polkadot/react-hooks';
 
-import useFutureSessionCommittee from '../Performance/useFutureSessionCommittee.js';
-import useSessionInfo from '../Performance/useSessionInfo.js';
-import ProducerPerformance from '../react-components/ProducerPerformance/index.js';
 import { useTranslation } from '../translate.js';
+import ValidatorFutureCommittees from './AlephCommitttee/ValidatorFutureCommittees.js';
 import ValidatorHistoricPerformance from './AlephCommitttee/ValidatorHistoricPerformance.js';
-import { range } from './util.js';
 import Validator from './Validator.js';
 
 interface Props {
@@ -40,9 +36,6 @@ function Query ({ className }: Props): React.ReactElement<Props> {
     [value]
   );
 
-  const sessionInfo = useSessionInfo();
-  const isNextTick = useNextTick();
-
   const groups = [
     { text: t('Past performance'), value: 'past' },
     { text: t('Future committees'), value: 'future' }
@@ -54,16 +47,6 @@ function Query ({ className }: Props): React.ReactElement<Props> {
   }, [api]
   );
 
-  const futureSessions = useMemo(() => {
-    if (sessionInfo) {
-      if (sessionInfo.currentSession < sessionInfo.maximumSessionNumber) {
-        return range(sessionInfo.maximumSessionNumber - sessionInfo.currentSession, sessionInfo.currentSession + 1);
-      }
-    }
-
-    return [];
-  }, [sessionInfo]);
-
   const eras = useCall<INumber[]>(api.derive.staking.erasHistoric);
 
   const labels = useMemo(
@@ -74,31 +57,6 @@ function Query ({ className }: Props): React.ReactElement<Props> {
   const _onQuery = useCallback(
     () => doQuery(validatorId),
     [validatorId]
-  );
-
-  const futureSessionCommittee = useFutureSessionCommittee(futureSessions);
-  const filteredSessionCommittee: FutureCommittee[] = useMemo(() => {
-    if (value) {
-      return futureSessionCommittee.filter((committee) => committee !== undefined && committee.producers.includes(value));
-    }
-
-    return [];
-  }, [futureSessionCommittee, value]);
-
-  const futureSessionsList: FutureCommittee[] = useMemo(
-    () => isNextTick
-      ? filteredSessionCommittee
-      : [],
-    [isNextTick, filteredSessionCommittee]
-  );
-
-  const headerRefFutureCommittee = useRef<[string, string, number?][]>(
-    [
-      [t('future committee sessions'), 'start', 1],
-      [t('session'), 'expand'],
-      [t('blocks created'), 'expand'],
-      [t('max % reward'), 'expand']
-    ]
   );
 
   if (!labels) {
@@ -142,27 +100,10 @@ function Query ({ className }: Props): React.ReactElement<Props> {
         />
       }
       {value && !!isAlephChain && groupIndex === 1 &&
-        <Table
-          className={className}
-          empty={filteredSessionCommittee.length === futureSessions.length && <div>{t('No entries found')}</div>}
-          emptySpinner={
-            <>
-              {(filteredSessionCommittee.length !== futureSessions.length) && <div>{t('Querying future sessions')}</div>}
-            </>
-          }
-          header={headerRefFutureCommittee.current}
-        >
-          {futureSessionsList?.map((committee): React.ReactNode => (
-            <ProducerPerformance
-              address={value}
-              blocksCreated={0}
-              filterName={''}
-              key={committee.session}
-              rewardPercentage={'0.0'}
-              session={committee.session}
-            />
-          ))}
-        </Table>}
+        <ValidatorFutureCommittees
+          address={value}
+        />
+      }
       {value && (
         <Validator
           labels={labels}
